requirements.txt
python-telegram-bot==20.6
apscheduler==3.11.0
httpx==0.25.2
tzlocal==5.3.1
